Redesign flaky test
