+++ frontend/editor/src/core/tests/stubbed/file-state-across-tools.spec.ts
@@ -9,6 +9,11 @@ const SAMPLE_PDF = path.join(FIXTURES_DIR, "sample.pdf");
  * Files uploaded on one tool page should remain in the workbench when the
  * user navigates to a different tool. This is FileContext behaviour and
  * easy to break with a stale-effect or unmount-clear bug.
+ *
+ * Covered for both navigation mechanisms, which take different code paths:
+ *   - a full page reload (page.goto) -> FileContext re-hydrates from IndexedDB
+ *   - an in-app tool-link click       -> client-side nav, FileContext stays
+ *                                        in memory
  */
 test.describe("File state persists across tool navigation", () => {
   test("file uploaded on /merge survives navigation to /split", async ({
@@ -30,6 +35,39 @@ test.describe("File state persists across tool navigation", () => {
     await page.getByTestId("my-files-button").click();
     await expect(page.getByText(/sample\.pdf/i).first()).toBeVisible({
       timeout: 5_000,
+    });
+  });
+
+  test("file uploaded on /merge survives an in-app tool-link navigation", async ({
+    page,
+  }) => {
+    await page.goto("/merge");
+    await page.waitForLoadState("domcontentloaded");
+    await uploadFiles(page, SAMPLE_PDF);
+
+    // Navigate via the in-app tool link (client-side React Router nav) rather
+    // than a full reload, so this exercises the in-memory FileContext path the
+    // page.goto test above doesn't. Fall back to a direct visit if the nav
+    // rail isn't showing the link yet.
+    const splitNav = page.getByRole("link", { name: /^Split$/i }).first();
+    if (await splitNav.isVisible({ timeout: 1_000 }).catch(() => false)) {
+      await splitNav.click();
+    } else {
+      await page.goto("/split");
+    }
+
+    // A client-side nav has no document load event, so waitForLoadState is a
+    // no-op here. Wait for the route to actually commit before opening the
+    // file manager; otherwise the my-files click fires mid-transition and
+    // opens it against a not-yet-settled workbench, which renders a permanent
+    // empty state (the flaky "0 items" that then passes on retry).
+    await expect(page).toHaveURL(/\/split(?:$|[/?#])/);
+
+    // The upload must still be listed after the tool switch. A "no files"
+    // empty state here would mean the client-side nav silently dropped it.
+    await page.getByTestId("my-files-button").click();
+    await expect(page.getByText(/sample\.pdf/i).first()).toBeVisible({
+      timeout: 10_000,
     });
   });
 });
