--- frontend/editor/src/core/tests/stubbed/unsaved-changes-guard.spec.ts
@@ -1,50 +1,0 @@
-import { test, expect } from "@app/tests/helpers/stub-test-base";
-import { uploadFiles } from "@app/tests/helpers/ui-helpers";
-import path from "path";
-
-const SAMPLE_PDF = path.join(__dirname, "../test-fixtures/sample.pdf");
-
-/**
- * The NavigationGuard context warns the user when they have unsaved work
- * (uploaded files or modified config) and try to navigate away. The guard
- * surface is a Mantine modal asking to confirm or cancel the navigation.
- *
- * Today the guard logic exists but is silently bypassed by tests that go
- * through the workbench. This spec asserts the modal appears and that
- * cancelling keeps the user on the current tool.
- */
-test.describe("Unsaved changes navigation guard", () => {
-  test("uploading then navigating away surfaces the guard prompt", async ({
-    page,
-  }) => {
-    await page.goto("/merge");
-    await page.waitForLoadState("domcontentloaded");
-    await uploadFiles(page, SAMPLE_PDF);
-
-    // Triggering a tool-level navigation while files are loaded should
-    // either prompt or clear-and-navigate cleanly. A regression that
-    // discards files silently is the failure we want to catch.
-    const splitNav = page.getByRole("link", { name: /^Split$/i }).first();
-    if (await splitNav.isVisible({ timeout: 1_000 }).catch(() => false)) {
-      await splitNav.click();
-    } else {
-      await page.goto("/split");
-    }
-    // Settle the navigation before opening My Files. The link-click branch
-    // routes client-side through the NavigationGuard, so without this the
-    // my-files click below can race the guard transition + IDB re-read.
-    await page.waitForLoadState("domcontentloaded");
-
-    // After arriving at /split the My Files page should still list the
-    // previously uploaded sample (NavigationGuard either kept us on
-    // /merge or moved us with state intact). A "no files" empty state
-    // here would indicate the guard silently dropped the workbench.
-    // Generous timeout: the file list is re-hydrated from IndexedDB after
-    // the cross-tool navigation, which can exceed 5s under --workers=3 CPU
-    // contention in CI (matches uploadFiles' own 10s IDB sync point).
-    await page.getByTestId("my-files-button").click();
-    await expect(page.getByText(/sample\.pdf/i).first()).toBeVisible({
-      timeout: 10_000,
-    });
-  });
-});
